makes memory a single input
Treating each address in the memory as a different input explodes the
state space as soon as we start reading, just reading a word already
creates 4^8=64k states.

Therefore, we will now treat memory as a single input, and treat the
memory value as a seed from which we derive a (potentially) different
value for each address.